+++ mrz-3-edition/tester.m
@@ -80,8 +80,13 @@ gaussCroutLuFactor([4 2 1; 9 8 7; 2 2 9]);  % [4 0 0; 9 3.5 0; 2 1 7.1429], [1 0
 choleskyFactorization([74 31 100; 31 54 89; 100 89 226]);  % [8.6023 0 0; 3.6037 6.4042 0; 11.6248 7.3558 6.0627]
 
 % exercises
-gaussPivotingWithPermutationSolver([2 1 3; 4 3 10; 2 4 17], [11 28 31]');  % example 4.3 in book with exercise 4.1 function
-exercise4_2()  % [0.1546 1.2336 1.9710 1.4386 1.6432]'
+gaussPivotingWithPermutationSolver([
+    2 1 3;
+    4 3 10;
+    2 4 17
+    ], [11 28 31]');  % example 4.3 in book with exercise 4.1 function
+% exercise4_2();  % [0.1546 1.2336 1.9710 1.4386 1.6432]'
+% exercise4_3();  % [7.5383 31.3009 -6.4753 -17.7043]'
 
 %% chapter 5
 % algorithms